Restructure into different source files
