+++ src/loader/mod.rs
@@ -1,0 +1,9 @@
+#[cfg(feature = "scene")]
+pub mod scene;
+#[cfg(feature = "scene")]
+pub use scene::*;
+
+#[cfg(not(feature = "scene"))]
+pub mod mesh;
+#[cfg(not(feature = "scene"))]
+pub use mesh::*;
