+++ src/loader/scene.rs
@@ -6,7 +6,6 @@ use bevy_render::{
 };
 use bevy_utils::BoxedFuture;
 use thiserror::Error;
-#[cfg(feature = "scene")]
 use {
     bevy_asset::AssetPath,
     bevy_ecs::world::{FromWorld, World},
@@ -21,11 +20,6 @@ use {
     std::path::PathBuf,
 };
 
-#[cfg(not(feature = "scene"))]
-#[derive(Default)]
-pub struct ObjLoader;
-
-#[cfg(feature = "scene")]
 pub struct ObjLoader {
     supported_compressed_formats: CompressedImageFormats,
 }
@@ -49,13 +43,7 @@ impl AssetLoader for ObjLoader {
         load_context: &'a mut LoadContext,
     ) -> BoxedFuture<'a, Result<(), anyhow::Error>> {
         Box::pin(async move {
-            Ok(load_obj(
-                bytes,
-                load_context,
-                #[cfg(feature = "scene")]
-                self.supported_compressed_formats,
-            )
-            .await?)
+            Ok(load_obj(bytes, load_context, self.supported_compressed_formats).await?)
         })
     }
 
@@ -65,7 +53,6 @@ impl AssetLoader for ObjLoader {
     }
 }
 
-#[cfg(feature = "scene")]
 impl FromWorld for ObjLoader {
     fn from_world(world: &mut World) -> Self {
         let supported_compressed_formats = match world.get_resource::<RenderDevice>() {
@@ -83,13 +70,10 @@ impl FromWorld for ObjLoader {
 pub enum ObjError {
     #[error("Invalid OBJ file: {0}")]
     TobjError(#[from] tobj::LoadError),
-    #[cfg(feature = "scene")]
     #[error("Invalid image file for texture: {0}")]
     InvalidImageFile(PathBuf),
-    #[cfg(feature = "scene")]
     #[error("Asset reading failed: {0}")]
     AssetIOError(#[from] bevy_asset::AssetIoError),
-    #[cfg(feature = "scene")]
     #[error("Texture conversion failed: {0}")]
     TextureError(#[from] bevy_render::texture::TextureError),
 }
@@ -97,18 +81,13 @@ pub enum ObjError {
 async fn load_obj<'a, 'b>(
     bytes: &'a [u8],
     load_context: &'a mut LoadContext<'b>,
-    #[cfg(feature = "scene")] supported_compressed_formats: CompressedImageFormats,
+    supported_compressed_formats: CompressedImageFormats,
 ) -> Result<(), ObjError> {
-    #[cfg(not(feature = "scene"))]
-    let obj = load_obj_from_bytes(bytes)?;
-    #[cfg(feature = "scene")]
     let obj = load_obj_from_bytes(bytes, load_context, supported_compressed_formats).await?;
-
     load_context.set_default_asset(LoadedAsset::new(obj));
     Ok(())
 }
 
-#[cfg(feature = "scene")]
 async fn load_texture_image<'a, 'b>(
     image_path: &'a str,
     load_context: &'a mut LoadContext<'b>,
@@ -132,7 +111,6 @@ async fn load_texture_image<'a, 'b>(
     )?)
 }
 
-#[cfg(feature = "scene")]
 async fn load_obj_from_bytes<'a, 'b>(
     mut bytes: &'a [u8],
     load_context: &'a mut LoadContext<'b>,
@@ -252,62 +230,4 @@ async fn load_obj_from_bytes<'a, 'b>(
     }
 
     Ok(Scene::new(world))
-}
-
-#[cfg(not(feature = "scene"))]
-pub fn load_obj_from_bytes(mut bytes: &[u8]) -> Result<Mesh, ObjError> {
-    let options = tobj::GPU_LOAD_OPTIONS;
-    let obj = tobj::load_obj_buf(&mut bytes, &options, |_m| {
-        Err(tobj::LoadError::OpenFileFailed)
-    })?;
-    let mut indices = Vec::new();
-    let mut vertex_position = Vec::new();
-    let mut vertex_normal = Vec::new();
-    let mut vertex_texture = Vec::new();
-    for model in obj.0 {
-        let index_offset = vertex_position.len() as u32; // Offset of the indices
-        indices.reserve(model.mesh.indices.len());
-        vertex_position.reserve(model.mesh.positions.len() / 3);
-        vertex_normal.reserve(model.mesh.normals.len() / 3);
-        vertex_texture.reserve(model.mesh.texcoords.len() / 2);
-        vertex_position.extend(
-            model
-                .mesh
-                .positions
-                .chunks_exact(3)
-                .map(|v| [v[0], v[1], v[2]]),
-        );
-        vertex_normal.extend(
-            model
-                .mesh
-                .normals
-                .chunks_exact(3)
-                .map(|n| [n[0], n[1], n[2]]),
-        );
-        vertex_texture.extend(
-            model
-                .mesh
-                .texcoords
-                .chunks_exact(2)
-                .map(|t| [t[0], 1.0 - t[1]]),
-        );
-        indices.extend(model.mesh.indices.iter().map(|i| i + index_offset));
-    }
-
-    let mut mesh = Mesh::new(PrimitiveTopology::TriangleList);
-    mesh.set_indices(Some(Indices::U32(indices)));
-
-    mesh.insert_attribute(Mesh::ATTRIBUTE_POSITION, vertex_position);
-    if !vertex_texture.is_empty() {
-        mesh.insert_attribute(Mesh::ATTRIBUTE_UV_0, vertex_texture);
-    }
-
-    if !vertex_normal.is_empty() {
-        mesh.insert_attribute(Mesh::ATTRIBUTE_NORMAL, vertex_normal);
-    } else {
-        mesh.duplicate_vertices();
-        mesh.compute_flat_normals();
-    }
-
-    Ok(mesh)
 }
